Kanbackend/requirements.txt
django-cors-headers==4.9.0
djangorestframework==3.16.1
Django==6.0.2
psycopg[binary]==3.3.2

# Tutorial defined packages
python-dotenv==1.2.1
